Rename files so they'll run in the correct order
