+++ test/01-Basic/01.first.test.server.js
@@ -1,8 +1,8 @@
 var expect = require('chai').expect;
 var request = require('request');
 
-var db = require('../server/config/db.js');
-var User = require('../server/models/user.js');
+var db = require('../../server/config/db.js');
+var User = require('../../server/models/user.js');
 
 var requestWithSession = request.defaults({jar: true});
 var serverURL = 'http://127.0.0.1:1337';
